+++ src/app/api/auth/[...nextauth]/route.ts
@@ -2,6 +2,4 @@ import { authConfig } from '@/app/auth' // Referring to the auth.ts we just crea
 import NextAuth from 'next-auth'
 
 const handler = NextAuth(authConfig)
-console.log(handler)
-
 export  { handler as GET, handler as POST }
