+++ src/app/components/cardComponents/interfaces.ts
@@ -1,0 +1,9 @@
+
+export interface CardsData {
+    title: string,
+    description: string,
+    imageSrc: string,
+    imageAlt: string,
+    width?: number,
+    height?: number,
+}
